refactor(#261): apply the phase1 review patterns to step-up and account
