+++ projects/app-web/src/routes/-account-2fa-verify/verify-two-factor-setup-handler.test.ts
@@ -2,38 +2,20 @@ import { expect, test } from 'bun:test';
 import { createId } from '@paralleldrive/cuid2';
 import { isRedirect } from '@tanstack/react-router';
 import { sessionCollection, userCollection, verificationCollection } from '../../mocks/db';
+import { buildFormData } from '../../test-utils/build-form-data';
 import { withRequestContext } from '../../test-utils/with-request-context';
 import { verifyTwoFactorSetupHandler } from './verify-two-factor-setup-handler';
 
-async function createSignedInUser(email: string): Promise<{
+async function createSignedInUser(): Promise<{
   readonly cookies: Readonly<Record<string, Readonly<Record<string, unknown>>>>;
   readonly userID: string;
 }> {
   const userID = createId();
   const sessionID = createId();
 
-  await userCollection.create({
-    createdAt: new Date(),
-    email,
-    id: userID,
-    name: 'Verify 2FA Setup',
-    password: 'password123',
-    seed: 0,
-    updatedAt: new Date(),
-    username: email.split('@')[0] ?? 'verify-2fa-setup',
-  });
+  await userCollection.create({ id: userID });
 
-  await sessionCollection.create({
-    createdAt: new Date(),
-    expiresAt: new Date(Date.now() + 60_000),
-    id: sessionID,
-    ipAddress: '127.0.0.1',
-    previousRefreshToken: null,
-    refreshToken: null,
-    updatedAt: new Date(),
-    userID,
-    verified: true,
-  });
+  await sessionCollection.create({ id: sessionID, userID });
 
   return {
     cookies: { en_session: { accessToken: sessionID, refreshToken: 'refresh', sessionID } },
@@ -41,18 +23,8 @@ async function createSignedInUser(email: string): Promise<{
   };
 }
 
-function buildFormData(fields: Readonly<Record<string, string>>): FormData {
-  const formData = new FormData();
-
-  for (const [key, value] of Object.entries(fields)) {
-    formData.set(key, value);
-  }
-
-  return formData;
-}
-
 test('it reports invalid code for a malformed submission', async () => {
-  const signedIn = await createSignedInUser('malformed-2fa-setup@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     verifyTwoFactorSetupHandler(buildFormData({ code: 'abc', target: signedIn.userID })),
@@ -62,10 +34,9 @@ test('it reports invalid code for a malformed submission', async () => {
 });
 
 test('it reports invalid code for an incorrect code', async () => {
-  const signedIn = await createSignedInUser('incorrect-2fa-setup@vers.test');
+  const signedIn = await createSignedInUser();
 
   await verificationCollection.create({
-    id: createId(),
     code: '654321',
     target: signedIn.userID,
     type: '2fa-setup',
@@ -82,14 +53,9 @@ test('it reports invalid code for an incorrect code', async () => {
 });
 
 test('it flips the verification to 2fa and redirects to account for a correct code', async () => {
-  const signedIn = await createSignedInUser('correct-2fa-setup@vers.test');
+  const signedIn = await createSignedInUser();
 
-  await verificationCollection.create({
-    id: createId(),
-    code: '123456',
-    target: signedIn.userID,
-    type: '2fa-setup',
-  });
+  await verificationCollection.create({ target: signedIn.userID, type: '2fa-setup' });
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, async () => {
     const redirectHref = await verifyTwoFactorSetupHandler(
