+++ projects/app-web/src/routes/-account-change-email/change-email-handler.test.ts
@@ -3,38 +3,20 @@ import { createId } from '@paralleldrive/cuid2';
 import { isRedirect } from '@tanstack/react-router';
 import { mintStepUpTransactionToken } from '../../lib/auth/step-up-transaction-token';
 import { sessionCollection, userCollection, verificationCollection } from '../../mocks/db';
+import { buildFormData } from '../../test-utils/build-form-data';
 import { withRequestContext } from '../../test-utils/with-request-context';
 import { changeEmailHandler } from './change-email-handler';
 
-async function createSignedInUser(email: string): Promise<{
+async function createSignedInUser(): Promise<{
   readonly cookies: Readonly<Record<string, Readonly<Record<string, unknown>>>>;
   readonly userID: string;
 }> {
   const userID = createId();
   const sessionID = createId();
 
-  await userCollection.create({
-    createdAt: new Date(),
-    email,
-    id: userID,
-    name: 'Change Email',
-    password: 'password123',
-    seed: 0,
-    updatedAt: new Date(),
-    username: email.split('@')[0] ?? 'change-email',
-  });
+  await userCollection.create({ id: userID });
 
-  await sessionCollection.create({
-    createdAt: new Date(),
-    expiresAt: new Date(Date.now() + 60_000),
-    id: sessionID,
-    ipAddress: '127.0.0.1',
-    previousRefreshToken: null,
-    refreshToken: null,
-    updatedAt: new Date(),
-    userID,
-    verified: true,
-  });
+  await sessionCollection.create({ id: sessionID, userID });
 
   return {
     cookies: { en_session: { accessToken: sessionID, refreshToken: 'refresh', sessionID } },
@@ -42,18 +24,8 @@ async function createSignedInUser(email: string): Promise<{
   };
 }
 
-function buildFormData(fields: Readonly<Record<string, string>>): FormData {
-  const formData = new FormData();
-
-  for (const [key, value] of Object.entries(fields)) {
-    formData.set(key, value);
-  }
-
-  return formData;
-}
-
 test('it reports a field error for an invalid email', async () => {
-  const signedIn = await createSignedInUser('invalid-email@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     changeEmailHandler(buildFormData({ email: 'not-an-email' })),
@@ -66,7 +38,7 @@ test('it reports a field error for an invalid email', async () => {
 });
 
 test('it starts a change-email verification and redirects to verify-otp for a caller with no 2FA', async () => {
-  const signedIn = await createSignedInUser('no-2fa-change@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, async () => {
     const redirectHref = await changeEmailHandler(buildFormData({ email: 'new@vers.test' }))
@@ -84,9 +56,9 @@ test('it starts a change-email verification and redirects to verify-otp for a ca
 });
 
 test('it reports step-up-required for a 2FA-enabled caller with no transaction token', async () => {
-  const signedIn = await createSignedInUser('gated-change@vers.test');
+  const signedIn = await createSignedInUser();
 
-  await verificationCollection.create({ id: createId(), target: signedIn.userID, type: '2fa' });
+  await verificationCollection.create({ target: signedIn.userID, type: '2fa' });
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     changeEmailHandler(buildFormData({ email: 'gated-new@vers.test' })),
@@ -96,9 +68,9 @@ test('it reports step-up-required for a 2FA-enabled caller with no transaction t
 });
 
 test('it applies the change once a valid step-up token is attached', async () => {
-  const signedIn = await createSignedInUser('token-change@vers.test');
+  const signedIn = await createSignedInUser();
 
-  await verificationCollection.create({ id: createId(), target: signedIn.userID, type: '2fa' });
+  await verificationCollection.create({ target: signedIn.userID, type: '2fa' });
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, async () => {
     const minted = await mintStepUpTransactionToken({
