+++ projects/app-web/src/routes/-account-change-password/change-password-form.test.tsx
@@ -4,12 +4,6 @@ import { renderWithRouter } from '../../test-utils/render-with-router';
 import { withRequestContext } from '../../test-utils/with-request-context';
 import { ChangePasswordForm } from './change-password-form';
 
-/**
- * `change-password-handler.test.ts` drives every `ChangePasswordResult` branch (field errors,
- * invalid credentials, step-up required, and the account redirect) against the handler body
- * directly — a plain result object never round-trips through an uncompiled `createServerFn`
- * export under `bun test`, so this file is limited to the initial render.
- */
 test('it renders the password fields and submit button', async () => {
   await withRequestContext({}, async () => {
     renderWithRouter(<ChangePasswordForm />);
