+++ projects/app-web/src/routes/-account-change-password/change-password-handler.test.ts
@@ -3,38 +3,20 @@ import { createId } from '@paralleldrive/cuid2';
 import { isRedirect } from '@tanstack/react-router';
 import { mintStepUpTransactionToken } from '../../lib/auth/step-up-transaction-token';
 import { sessionCollection, userCollection, verificationCollection } from '../../mocks/db';
+import { buildFormData } from '../../test-utils/build-form-data';
 import { withRequestContext } from '../../test-utils/with-request-context';
 import { changePasswordHandler } from './change-password-handler';
 
-async function createSignedInUser(email: string): Promise<{
+async function createSignedInUser(): Promise<{
   readonly cookies: Readonly<Record<string, Readonly<Record<string, unknown>>>>;
   readonly userID: string;
 }> {
   const userID = createId();
   const sessionID = createId();
 
-  await userCollection.create({
-    createdAt: new Date(),
-    email,
-    id: userID,
-    name: 'Change Password',
-    password: 'original-password',
-    seed: 0,
-    updatedAt: new Date(),
-    username: email.split('@')[0] ?? 'change-password',
-  });
+  await userCollection.create({ id: userID, password: 'original-password' });
 
-  await sessionCollection.create({
-    createdAt: new Date(),
-    expiresAt: new Date(Date.now() + 60_000),
-    id: sessionID,
-    ipAddress: '127.0.0.1',
-    previousRefreshToken: null,
-    refreshToken: null,
-    updatedAt: new Date(),
-    userID,
-    verified: true,
-  });
+  await sessionCollection.create({ id: sessionID, userID });
 
   return {
     cookies: { en_session: { accessToken: sessionID, refreshToken: 'refresh', sessionID } },
@@ -42,20 +24,10 @@ async function createSignedInUser(email: string): Promise<{
   };
 }
 
-function buildFormData(fields: Readonly<Record<string, string>>): FormData {
-  const formData = new FormData();
-
-  for (const [key, value] of Object.entries(fields)) {
-    formData.set(key, value);
-  }
-
-  return formData;
-}
-
 const validNewPassword = { confirmPassword: 'new-password123', password: 'new-password123' };
 
 test('it reports a field error when the new passwords do not match', async () => {
-  const signedIn = await createSignedInUser('mismatch@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     changePasswordHandler(
@@ -74,7 +46,7 @@ test('it reports a field error when the new passwords do not match', async () =>
 });
 
 test('it reports invalid credentials for the wrong current password', async () => {
-  const signedIn = await createSignedInUser('wrong-current@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     changePasswordHandler(
@@ -89,7 +61,7 @@ test('it reports invalid credentials for the wrong current password', async () =
 });
 
 test('it changes the password and redirects to account for a caller with no 2FA', async () => {
-  const signedIn = await createSignedInUser('no-2fa-password@vers.test');
+  const signedIn = await createSignedInUser();
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, async () => {
     const redirectHref = await changePasswordHandler(
@@ -109,9 +81,9 @@ test('it changes the password and redirects to account for a caller with no 2FA'
 });
 
 test('it reports step-up-required for a 2FA-enabled caller with no transaction token', async () => {
-  const signedIn = await createSignedInUser('gated-password@vers.test');
+  const signedIn = await createSignedInUser();
 
-  await verificationCollection.create({ id: createId(), target: signedIn.userID, type: '2fa' });
+  await verificationCollection.create({ target: signedIn.userID, type: '2fa' });
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, () =>
     changePasswordHandler(
@@ -123,9 +95,9 @@ test('it reports step-up-required for a 2FA-enabled caller with no transaction t
 });
 
 test('it changes the password once a valid step-up token is attached', async () => {
-  const signedIn = await createSignedInUser('token-password@vers.test');
+  const signedIn = await createSignedInUser();
 
-  await verificationCollection.create({ id: createId(), target: signedIn.userID, type: '2fa' });
+  await verificationCollection.create({ target: signedIn.userID, type: '2fa' });
 
   const outcome = await withRequestContext({ cookies: signedIn.cookies }, async () => {
     const minted = await mintStepUpTransactionToken({
